@@ -24,7 +24,7 @@ import { map } from 'rxjs';
 })
 export class BlogHolderComponent {
   blogs$ = this.sts.useScullyTransferState(
-    'blogRotues',
+    'blogRoutes',
     this.srs.available$.pipe(
       map((routeList) => routeList.filter((route: ScullyRoute) => route.route.startsWith(`/blog/`))),
       map((blogs) => blogs.sort((a, b) => (a.date < b.date ? -1 : 1)))
