@@ -16,7 +16,6 @@ export class StaticComponent implements OnInit {
   topLevel$ = this.srs.topLevel$;
 
   title$ = this.srs.getCurrent().pipe(
-    tap(r => console.log('current route', r)),
     map(r => r.title || ''),
     tap(t => this.title.setTitle(t))
   );
@@ -24,14 +23,14 @@ export class StaticComponent implements OnInit {
     private srs: ScullyRoutesService,
     private route: ActivatedRoute,
     private title: Title
-  ) {}
+  ) { }
 
   get routes() {
     return this.unPublished
       ? this.srs.unPublished$
       : this.toplevelOnly
-      ? this.topLevel$
-      : this.available$;
+        ? this.topLevel$
+        : this.available$;
   }
 
   ngOnInit() {
