@@ -1,3 +1,4 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { NavigationEnd, Router } from '@angular/router';
 import {
@@ -27,7 +28,7 @@ export class ScullyRoutesService {
    * An observable with all routes, published and unpublished alike
    */
   allRoutes$: Observable<ScullyRoute[]> = this.refresh.pipe(
-    switchMap(() => fetchHttp<ScullyRoute[]>('assets/scully-routes.json')),
+    switchMap(() => this.http.get('assets/scully-routes.json')),
     catchError(() => {
       console.warn(
         'Scully routes file not found, are you running the Scully generated version of your site?'
@@ -74,7 +75,7 @@ export class ScullyRoutesService {
     shareReplay({ refCount: false, bufferSize: 1 })
   );
 
-  constructor(private router: Router) {
+  constructor(private router: Router, private http: HttpClient) {
     /** kick off first cycle */
     this.reload();
   }
