@@ -27,9 +27,9 @@ import { routeSplit } from './lib/utils/routeSplit';
 import { staticServer } from './lib/utils/serverstuff/staticServer';
 import { getHandledRoutes } from './lib/utils/services/routeStorage';
 import { startScully } from './lib/utils/startup';
-import { universalRender } from './lib/systemPlugins/universalRender';
 export * from './lib/utils/log';
 export * from './lib/utils/procesmanager';
+export * from './lib/utils/platform-server'
 export {
   configValidator,
   ContentMetaData,
@@ -49,7 +49,6 @@ export {
   ScullyConfig,
   setPluginConfig,
   startScully,
-  universalRender,
   updateScullyConfig,
   /** WIP part, those might be remove again in near future. */
   handleTravesal,
