@@ -1,16 +1,19 @@
 // const asciidoctor = require('asciidoctor.js')();
 import { registerPlugin } from '../pluginManagement/pluginRepository';
 import { logError } from '../utils';
+import Processor from 'asciidoctor';
+
+const asciiDoctor = Processor();
 
 const asciiDocPlugin = async (raw: string) => {
   try {
-    const asciidoctor: any = require('asciidoctor.js')();
-    return asciidoctor.convert(raw, {});
+    const html = asciiDoctor.convert(raw, { safe: 'unsafe' });
+    return html as unknown as string;
   } catch (e) {
     logError(
-      `When you want to use asciidoc, you need to install the optional asciidocor.js module by:
+      `When you want to use asciidoc, you need to install the optional asciidocor module by:
 
-  npm install -s asciidoctor.js
+  npm install -s asciidoctor
 
 `,
       e
@@ -19,3 +22,12 @@ const asciiDocPlugin = async (raw: string) => {
 };
 
 registerPlugin('fileHandler', 'adoc', asciiDocPlugin, ['asciidoc', 'asc']);
+
+
+// {
+//   runtime: {
+//     platform: 'node',
+//     engine: 'v8',
+//     framework: 'webworker'
+//   }
+// }
