@@ -1,5 +1,7 @@
-export * from './content-render-utils';
 export * from './contentRenderPlugin';
+export * from './contentTextRenderPlugin';
+export * from './executePlugins';
+export * from './jsdomPlugins';
 export * from './launchedBrowser';
 export * from './puppeteerRenderPlugin';
 export * from './playwrightRenderPlugin';
