@@ -16,6 +16,8 @@ export interface RouteConfig {
 }
 
 export interface HandledRoute {
+  /** the string as used in the Scully config */
+  usedConfigRoute?:string;
   /** the _complete_ route */
   route: string;
   /** the raw route, will be used by puppeteer over the route.route, will be used as is. must include the http(s):// part and eventual params*/
@@ -66,5 +68,5 @@ export interface ContentTextRoute extends HandledRoute {
   /** the type of content (MD/HTML this determines what plugin us used for render) */
   contentType?: string;
   /** The actual raw content that will be rendered into scully-content */
-  content?: string;
+  content?: string|((route?:HandledRoute)=>string);
 }
