@@ -5,11 +5,11 @@ import { HandledRoute } from '../routerPlugins/handledRoute.interface';
 import { watch } from '../utils/cli-options';
 import { scullyConfig } from '../utils/config';
 import { createFolderFor } from '../utils/createFolderFor';
-import { log, logError, printProgress, yellow } from '../utils/log';
+import { log, logError, logOk, printProgress, yellow } from '../utils/log';
 
 export const routesFileName = '/assets/scully-routes.json';
 
-export const storeRoutes = Symbol('storeRoutes');
+export const storeRoutes = 'storeRoutes' as const;
 registerPlugin(scullySystem, storeRoutes, storeRoutesPlugin);
 async function storeRoutesPlugin(routes: HandledRoute[]) {
   /** in the angular source folder */
@@ -35,7 +35,7 @@ async function storeRoutesPlugin(routes: HandledRoute[]) {
         const existing = readFileSync(srcFile).toString('utf-8').trim();
         if (jsonResult.trim() === existing) {
           /** the same. done, don't write */
-          log('keep existing route file');
+          logOk('keep existing route file');
           return;
         }
       } catch (e) {
@@ -48,9 +48,9 @@ async function storeRoutesPlugin(routes: HandledRoute[]) {
       writeFileSync(file, jsonResult);
     };
     files.forEach(write);
-    log(`Route list created in files:${files.map(
+    logOk(`Route list created in files:${files.map(
       (f) => `
-  "${yellow(f)}"`
+     "${yellow(f)}"`
     )}
 `);
   } catch {
