@@ -1,13 +1,14 @@
 import { performance } from 'perf_hooks';
+import { registerPlugin, scullySystem } from '../../pluginManagement';
 import { findPlugin } from '../../pluginManagement/pluginConfig';
 import { traverseAppRoutes } from '../../routerPlugins/traverseAppRoutesPlugin';
 import { flushRawRoutesCache, rawRoutesCache } from '../cache';
-import { log, logWarn } from '../log';
+import { logOk, logWarn, printProgress } from '../log';
 import { performanceIds } from '../performanceIds';
-import { registerPlugin, scullySystem } from '../../pluginManagement';
 
-export const handleTravesal = Symbol('handleTravesal');
+export const handleTravesal = 'handleTravesal' as const;
 registerPlugin(scullySystem, handleTravesal, plugin);
+let once = true
 
 async function plugin({ forceScan } = { forceScan: false }): Promise<string[]> {
   let unhandledRoutes: string[];
@@ -15,7 +16,8 @@ async function plugin({ forceScan } = { forceScan: false }): Promise<string[]> {
     flushRawRoutesCache();
   }
   if (rawRoutesCache.size === 0) {
-    log('Finding all routes in application.');
+    printProgress(undefined, 'Finding routes in app');
+    const start = performance.now();
     performance.mark('startTraverse');
     unhandledRoutes = await findPlugin(traverseAppRoutes)();
     performance.mark('stopTraverse');
@@ -23,6 +25,10 @@ async function plugin({ forceScan } = { forceScan: false }): Promise<string[]> {
     unhandledRoutes.forEach((r) => rawRoutesCache.add(r));
   } else {
     unhandledRoutes = [...rawRoutesCache.keys()];
+    if (once) {
+      once = false;
+      logOk(`Successfully retrieved routes from cache`);
+    }
   }
   if (unhandledRoutes.length < 1) {
     logWarn('No routes found in application, are you sure you installed the router? Terminating.');
