@@ -1,9 +1,13 @@
 export * from './handlers';
+export * from './platform-server';
+export * from './procesmanager';
 export * from './serverstuff';
 export * from './services';
+export * from './startup';
 export * from './assertNever';
 export * from './asyncPool';
 export * from './cache';
+export * from './captureMessage';
 export * from './chunk';
 export * from './cli-options';
 export * from './compileConfig';
@@ -11,6 +15,7 @@ export * from './config';
 export * from './createFolderFor';
 export * from './deepClone';
 export * from './deepGet';
+export * from './exitHandler';
 export * from './findAngularJsonPath';
 export * from './fsAngular';
 export * from './fsFolder';
@@ -24,6 +29,5 @@ export * from './read-angular-json';
 export * from './replaceFirstRouteParamWithVal';
 export * from './restart';
 export * from './routeSplit';
-export * from './startup';
+export * from './scullydot';
 export * from './validateConfig';
-export * from './procesmanager'
