@@ -3,7 +3,9 @@ import { logError } from '../log';
 import { Tasks } from './tasks.interface';
 
 const masterTaskList: Tasks = {
-  kill: () => process.exit(0),
+  kill: () => {
+    process.exit(0)
+  },
 };
 const workerMessages = new Subject<{ type: string; msg: any }>();
 export const workerMessages$ = workerMessages.asObservable();
@@ -18,12 +20,19 @@ export const addWorkerTask = (key: string, fn: (...msg: any[]) => any | Promise<
 
 export function startWorkerListener(tasks: Tasks) {
   /** check if I'm running in an forked task, will be undefined otherwise */
-  if (process.send) {
+  if (process.env.SCULLY_WORKER === 'true') {
+    let watchDog: NodeJS.Timeout;
     /** add tasks to master task list. */
     Object.entries(tasks).forEach(([key, task]) => addWorkerTask(key, task));
 
     process.on('message', async ([type, msg]: [string, any]) => {
-      // console.log('got msg',type,msg)
+      watchDog && clearTimeout(watchDog);
+      //TODO: do we need to make the 5 minute timeout configurable?
+      watchDog = setTimeout(() => {
+        /** terminate worker after 5 minutes of inactivity */
+        process.exit(0);
+      }, 5 * 60 * 1000);
+      // console.log('got msg',type,msg)s
       if (masterTaskList.hasOwnProperty(type)) {
         try {
           const result = await masterTaskList[type](msg);
@@ -37,6 +46,15 @@ export function startWorkerListener(tasks: Tasks) {
         workerMessages.next({ type, msg });
       }
     });
+
+    process.on('exit', () => {
+      process.exit(0);
+    })
+
+    process.on('uncaughtException', (err: Error) => {
+      logError(`Uncaught exception in worker thread ${err}`);
+      process.exit(15);
+    })
 
     process.send('ready');
     return;
